modify EchoData destructor so that it only creates a Path object, if the temp_zarr_dir exists

+++ echopype/echodata/echodata.py
@@ -84,13 +84,15 @@ class EchoData:
         # TODO: this destructor seems to not work in Jupyter Lab if restart or
         #  even clear all outputs is used. It will work if you explicitly delete the object
 
-        # get Path object of temporary zarr directory created by Parsed2Zarr
-        p2z_temp_dir = Path(self.parsed2zarr_obj.temp_zarr_dir)
+        if self.parsed2zarr_obj.temp_zarr_dir is not None:
 
-        # remove temporary directory created by Parsed2Zarr, if it exists
-        if p2z_temp_dir.exists():
-            # TODO: do we need to check file permissions here?
-            shutil.rmtree(p2z_temp_dir)
+            # get Path object of temporary zarr directory created by Parsed2Zarr
+            p2z_temp_dir = Path(self.parsed2zarr_obj.temp_zarr_dir)
+
+            # remove temporary directory created by Parsed2Zarr, if it exists
+            if p2z_temp_dir.exists():
+                # TODO: do we need to check file permissions here?
+                shutil.rmtree(p2z_temp_dir)
 
     def __str__(self) -> str:
         fpath = "Internal Memory"
